GET-326 Move Job Profile search out of model into its own class with tests
jira-ticket: https://dfedigital.atlassian.net/browse/GET-326

* Add validations for presence of query term
* Return no records if validation fails
