+++ app/models/job_profile.rb
@@ -17,34 +17,6 @@ class JobProfile < ApplicationRecord
     where('lower(name) = ?', name.downcase).first
   end
 
-  def self.search(string)
-    return none unless string.present?
-
-    query_string = QueryStringFormatter.format(string)
-
-    select(:id, :name, :description, :alternative_titles, :slug, :salary_min, :salary_max)
-      .where(build_text_search_query, query_string: query_string)
-      .order(
-        Arel.sql(build_rank_query(query_string)) => :desc
-      )
-  end
-
-  def self.build_text_search_query
-    <<-SQL
-      to_tsvector('english', name) @@ to_tsquery('english', :query_string)
-      OR to_tsvector('english', alternative_titles) @@ to_tsquery('english', :query_string)
-      OR to_tsvector('english', description) @@ to_tsquery('english', :query_string)
-    SQL
-  end
-
-  def self.build_rank_query(query_string)
-    quoted_query_string = ActiveRecord::Base.connection.quote(query_string)
-
-    <<-RANK
-      ts_rank(to_tsvector('english', name), to_tsquery('english', #{quoted_query_string}))
-    RANK
-  end
-
   def self.import(slug, url)
     find_or_create_by(slug: slug) do |job_profile|
       job_profile.update(source_url: url, name: slug.titleize)
