+++ spec/models/job_profile_search_spec.rb
@@ -1,0 +1,142 @@
+require 'rails_helper'
+
+RSpec.describe JobProfileSearch do
+  describe '.search' do
+    it 'returns a job profile if a name has accidental spaces' do
+      job_profile = create(:job_profile, name: 'Music therapist')
+
+      expect(described_class.new(term: 'Music    therapist      ').search).to contain_exactly(job_profile)
+    end
+
+    it 'returns a job profile if a name matches exactly' do
+      job_profile = create(:job_profile, name: 'Beverage Dissemination Officer')
+
+      expect(described_class.new(term: 'Beverage Dissemination Officer').search).to contain_exactly(job_profile)
+    end
+
+    it 'returns a job profile if a name is like an existing job profile name' do
+      job_profile = create(:job_profile, name: 'Beverage Dissemination Officer')
+
+      expect(described_class.new(term: 'Dissemination').search).to contain_exactly(job_profile)
+    end
+
+    it 'returns a job profile if a name supplied is uppercase' do
+      job_profile = create(:job_profile, name: 'Beverage Dissemination Officer')
+
+      expect(described_class.new(term: 'DISSEMINATION').search).to contain_exactly(job_profile)
+    end
+
+    it 'returns a job profile if a name supplied is lowercase' do
+      job_profile = create(:job_profile, name: 'BEVERAGE DISSEMINATION OFFICER')
+
+      expect(described_class.new(term: 'beverage dissemination officer').search).to contain_exactly(job_profile)
+    end
+
+    it 'returns a job profile if a name supplied is mixed case' do
+      job_profile = create(:job_profile, name: 'Beverage Dissemination Officer')
+
+      expect(described_class.new(term: 'Beverage Dissemination OfFicer').search).to contain_exactly(job_profile)
+    end
+
+    it 'returns nothing if no job profile is matched' do
+      create(:job_profile, name: 'Dream Alchemist')
+
+      expect(described_class.new(term: 'Beverage Dissemination Officer').search).to be_empty
+    end
+
+    it 'returns no records if empty name is supplied' do
+      create(:job_profile, name: 'Dream Alchemist')
+      create(:job_profile, name: 'Beverage Dissemination Officer')
+
+      expect(described_class.new(term: '').search).to be_empty
+    end
+
+    it 'returns no records if no name is supplied' do
+      create(:job_profile, name: 'Dream Alchemist')
+
+      expect(described_class.new(term: nil).search).to be_empty
+    end
+
+    it 'returns a record if value is in alternative titles only' do
+      job_profile = create(:job_profile, name: 'Chef', alternative_titles: 'Cook, Kitchen manager')
+
+      expect(described_class.new(term: 'cook').search).to contain_exactly(job_profile)
+    end
+
+    it 'returns a record if value is in the description only' do
+      job_profile = create(:job_profile, name: 'Chef', alternative_titles: 'Cook', description: 'Kitchen manager')
+
+      expect(described_class.new(term: 'Kitchen manager').search).to contain_exactly(job_profile)
+    end
+
+    it 'matches different records in query' do
+      job_profiles = [
+        create(:job_profile, name: 'Head Chef'),
+        create(:job_profile, alternative_titles: 'Cook'),
+        create(:job_profile, description: 'Street food traders ')
+      ]
+
+      expect(described_class.new(term: 'food cook chef').search).to eq(job_profiles)
+    end
+
+    it 'matches query words as prefixes to records' do
+      job_profile = create(:job_profile, name: 'Chief executive')
+
+      expect(described_class.new(term: 'exec').search).to contain_exactly(job_profile)
+    end
+
+    it 'matches query words through stemming to records' do
+      job_profile = create(:job_profile, name: 'Chief executive')
+
+      expect(described_class.new(term: 'chiefs').search).to contain_exactly(job_profile)
+    end
+
+    it 'ignores unicode characters in query string' do
+      job_profile = create(:job_profile, name: 'Chief executive')
+
+      expect(described_class.new(term: "ṩ ® chief's àⅣ <>; /").search).to contain_exactly(job_profile)
+    end
+
+    it 'ignores irrelevant characters in query string' do
+      job_profile = create(:job_profile, name: 'Script Writer')
+
+      expect(described_class.new(term: '<script> alert("PWNED!") </script>').search).to contain_exactly(job_profile)
+    end
+
+    it 'guards against injection in string' do
+      job_profile = create(:job_profile, name: 'Table Writer')
+
+      expect(described_class.new(term: 'name; drop table job_profiles;').search).to contain_exactly(job_profile)
+    end
+
+    it 'orders records according to matched name' do
+      cook = create(:job_profile, alternative_titles: 'Cook')
+      head_chef = create(:job_profile, name: 'Head Chef')
+      food_trader = create(:job_profile, description: 'Street food traders ')
+      kitchen_chef = create(:job_profile, name: 'Kitchen Chef')
+
+      job_profiles_in_order = [
+        head_chef,
+        kitchen_chef,
+        cook,
+        food_trader
+      ]
+
+      expect(described_class.new(term: 'food cook chef').search).to eq(job_profiles_in_order)
+    end
+  end
+
+  describe 'validation' do
+    it 'is invalid if term is not entered' do
+      search = described_class.new(term: nil)
+
+      expect(search).not_to be_valid
+    end
+
+    it 'is invalid if empty term entered' do
+      search = described_class.new(term: '')
+
+      expect(search).not_to be_valid
+    end
+  end
+end
